// generated/src/app/api-models/cm-adjustment.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CmAdjustment {
  seqCmadjId: number;
  seqAgentId: number;
  agentId: string;
  customerType: string;
  customerId: string;
  transactionDate: Date;
  transactionReason: string;
  transactionAmt: number;
  billingMonth: Date;
  adjustmentStatus: string;
  invoiceNumber: number;
  postDate: Date;
  companyCode: string;
  glRefCode: string;
  transactionDesc: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}